[Add]: Added submodule `berkley-softfloat-3` for myos

+++ i_video.h
@@ -148,7 +148,7 @@ struct color
 extern char *video_driver;
 extern boolean screenvisible;
 
-extern float mouse_acceleration;
+extern int mouse_acceleration;
 extern int mouse_threshold;
 extern int vanilla_keyboard_mapping;
 extern boolean screensaver_mode;
